% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async_utils.R
\name{async_data_aggregation}
\alias{async_data_aggregation}
\title{Process data aggregation asynchronously}
\usage{
async_data_aggregation(
  data,
  group_vars,
  summary_vars,
  label = "Data aggregation"
)
}
\arguments{
\item{data}{Input data}

\item{group_vars}{Grouping variables}

\item{summary_vars}{Summary variables}

\item{label}{Label for logging}
}
\description{
Process data aggregation asynchronously
}
