% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async_utils.R
\name{background_processor}
\alias{background_processor}
\title{Create a background data processor}
\usage{
background_processor(
  data_source,
  processing_fn,
  update_interval_ms = 5000,
  label = "Background processor"
)
}
\arguments{
\item{data_source}{Reactive data source}

\item{processing_fn}{Processing function}

\item{update_interval_ms}{How often to check for updates}

\item{label}{Label for logging}
}
\description{
Create a background data processor
}
