% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_utils.R
\name{batch_reactives}
\alias{batch_reactives}
\title{Batch multiple reactive updates}
\usage{
batch_reactives(..., batch_size = 3)
}
\arguments{
\item{...}{Named reactive expressions}

\item{batch_size}{Number of updates to batch together}
}
\description{
Batch multiple reactive updates
}
