% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_optimization.R
\name{compress_data}
\alias{compress_data}
\title{Data serialization and transfer optimization utilities}
\usage{
compress_data(data, method = "gzip")
}
\arguments{
\item{data}{Data to compress}

\item{method}{Compression method ("gzip", "bzip2", "xz")}
}
\description{
Utilities for optimizing data transfer between server and client
}
