% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_data_last_updated.R
\docType{data}
\name{data_last_updated}
\alias{data_last_updated}
\title{Last updated from peskas}
\format{
A dataframe
}
\usage{
data_last_updated
}
\description{
Last updated from peskas
}
\keyword{datasets}
