% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_optimization.R
\name{decompress_data}
\alias{decompress_data}
\title{Decompress data}
\usage{
decompress_data(compressed_data)
}
\arguments{
\item{compressed_data}{Compressed data object}
}
\description{
Decompress data
}
