% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{dummy_chart}
\alias{dummy_chart}
\title{Dummy apex chart}
\usage{
dummy_chart(type = "bar", sparkline.enabled = TRUE)
}
\arguments{
\item{type}{Plot type}

\item{sparkline.enabled}{Enable sparkline}
}
\value{
an htmlwidget
}
\description{
Dummy apex chart
}
