% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_utils.R
\name{init_performance_monitoring}
\alias{init_performance_monitoring}
\title{Initialize performance monitoring}
\usage{
init_performance_monitoring(session)
}
\arguments{
\item{session}{Shiny session object}
}
\description{
Initialize performance monitoring
}
