% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_footer.R
\name{inline_li_link}
\alias{inline_li_link}
\title{Creates a inline list item with a link}
\usage{
inline_li_link(content = "Link text", href = "#", target = "_blank")
}
\value{
a shiny tag "li" element
}
\description{
Te be used in elements like the footer
}
