% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_optimization.R
\name{lazy_data_manager}
\alias{lazy_data_manager}
\title{Lazy loading data manager}
\usage{
lazy_data_manager(data_source, cache_duration_seconds = 300)
}
\arguments{
\item{data_source}{Function that returns data}

\item{cache_duration_seconds}{How long to cache data}
}
\description{
Lazy loading data manager
}
