% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_language.R
\name{mod_language_ui}
\alias{mod_language_ui}
\title{language UI Function}
\usage{
mod_language_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
