% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_utils.R
\name{monitor_reactive}
\alias{monitor_reactive}
\title{Monitor reactive performance}
\usage{
monitor_reactive(reactive_expr, label = "Reactive")
}
\arguments{
\item{reactive_expr}{Reactive expression}

\item{label}{Label for monitoring}
}
\description{
Monitor reactive performance
}
