% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_municipal_aggregated.R
\docType{data}
\name{municipal_aggregated}
\alias{municipal_aggregated}
\title{Municipal data aggregated}
\format{
A dataframe
}
\usage{
municipal_aggregated
}
\description{
Municipal data aggregated
}
\keyword{datasets}
