% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_municipal_taxa.R
\docType{data}
\name{municipal_taxa}
\alias{municipal_taxa}
\title{Municipal taxa data aggregated}
\format{
A dataframe
}
\usage{
municipal_taxa
}
\description{
Municipal taxa data aggregated
}
\keyword{datasets}
