% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nutrients_aggregated.R
\docType{data}
\name{nutrients_aggregated}
\alias{nutrients_aggregated}
\title{Nutrients data aggregated}
\format{
A dataframe
}
\usage{
nutrients_aggregated
}
\description{
Nutrients data aggregated
}
\keyword{datasets}
