% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_optimization.R
\name{optimize_chart_data}
\alias{optimize_chart_data}
\title{Optimize JSON output for charts}
\usage{
optimize_chart_data(data, precision = 3)
}
\arguments{
\item{data}{Chart data}

\item{precision}{Numeric precision}
}
\description{
Optimize JSON output for charts
}
