% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_optimization.R
\name{optimize_dataframe}
\alias{optimize_dataframe}
\title{Optimize data frame for transfer}
\usage{
optimize_dataframe(df, precision = 2, max_rows = 10000)
}
\arguments{
\item{df}{Data frame to optimize}

\item{precision}{Number of decimal places for numeric columns}

\item{max_rows}{Maximum number of rows to include}
}
\description{
Optimize data frame for transfer
}
