% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_page_heading.R
\name{page_heading}
\alias{page_heading}
\title{Page heading for the}
\usage{
page_heading(
  pretitle = "Page pretitle",
  title = "Page title",
  download_text = "Download text",
  ...
)
}
\arguments{
\item{pretitle}{Text to appear on top of title}

\item{title}{Text to appear as the title}

\item{...}{Elements to add at the right hand side of the title (e.g. buttons, etc)}
}
\value{
a shiny tag
}
\description{
Page heading for the
}
