% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_optimization.R
\name{paginate_data}
\alias{paginate_data}
\title{Create paginated data response}
\usage{
paginate_data(data, page = 1, page_size = 100)
}
\arguments{
\item{data}{Full dataset}

\item{page}{Page number (1-based)}

\item{page_size}{Number of records per page}
}
\description{
Create paginated data response
}
