% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_horizontal_barplot}
\alias{plot_horizontal_barplot}
\title{Plot series using apexcharter}
\usage{
plot_horizontal_barplot(
  x_categories,
  series,
  y_formatter,
  x_formatter,
  type = "bar",
  sparkline = F,
  colors = NULL
)
}
\arguments{
\item{x_categories}{X axis categories}

\item{series}{Time series in list format}

\item{y_formatter}{Y axis formatter}

\item{x_formatter}{X axis formatter}

\item{type}{Plot type}

\item{sparkline}{Enable sparkiline}

\item{colors}{Plot colors}
}
\value{
An apexcharter barplot
}
\description{
Plot series using apexcharter
}
