% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot series using apexcharter}
\usage{
plot_timeseries(x_categories, series, y_formatter, type = "bar", sparkline = F)
}
\description{
Plot series using apexcharter
}
