% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries_apex}
\alias{plot_timeseries_apex}
\title{Plot series using apexcharter for more custom use}
\usage{
plot_timeseries_apex(
  data = NULL,
  type = "area",
  sparkline = F,
  y_formatter = NULL,
  plot_color = "#206bc4",
  lab_color = "#000000",
  mean = FALSE
)
}
\description{
Plot series using apexcharter for more custom use
}
