% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async_utils.R
\name{progressive_module_loader}
\alias{progressive_module_loader}
\title{Create a progressive loader for heavy modules}
\usage{
progressive_module_loader(module_loaders, load_delay_ms = 500)
}
\arguments{
\item{module_loaders}{List of module loading functions}

\item{load_delay_ms}{Delay between loading each module}
}
\description{
Create a progressive loader for heavy modules
}
