% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_boundary.R
\name{resilient_reactive}
\alias{resilient_reactive}
\title{Resilient reactive expression}
\usage{
resilient_reactive(
  expr,
  fallback_value = NULL,
  retry_attempts = 3,
  retry_delay_ms = 1000,
  label = "Resilient reactive"
)
}
\arguments{
\item{expr}{Reactive expression}

\item{fallback_value}{Value to return on error}

\item{retry_attempts}{Number of retry attempts}

\item{retry_delay_ms}{Delay between retries in milliseconds}

\item{label}{Label for logging}
}
\description{
Resilient reactive expression
}
