% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_boundary.R
\name{safe_module_server}
\alias{safe_module_server}
\title{Safe module server wrapper}
\usage{
safe_module_server(module_server, id, ..., label = paste("Module", id))
}
\arguments{
\item{module_server}{Module server function}

\item{id}{Module ID}

\item{...}{Additional arguments to pass to module server}

\item{label}{Label for logging}
}
\description{
Safe module server wrapper
}
