% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_utils.R
\name{safe_reactive}
\alias{safe_reactive}
\title{Reactive utilities for safe and performant reactive programming}
\usage{
safe_reactive(
  expr,
  validators = list(),
  fallback_value = NULL,
  label = "Safe reactive"
)
}
\arguments{
\item{expr}{Reactive expression}

\item{validators}{List of validation functions}

\item{fallback_value}{Value to return if validation fails}

\item{label}{Label for logging}
}
\description{
Utilities for creating safer and more performant reactive expressions
}
