% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async_utils.R
\name{setup_async_environment}
\alias{setup_async_environment}
\title{Setup async processing environment}
\usage{
setup_async_environment(max_workers = 2)
}
\arguments{
\item{max_workers}{Maximum number of background workers}
}
\description{
Setup async processing environment
}
