% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_boundary.R
\name{setup_health_monitor}
\alias{setup_health_monitor}
\title{Create a health check system}
\usage{
setup_health_monitor(checks, check_interval_ms = 30000)
}
\arguments{
\item{checks}{List of health check functions}

\item{check_interval_ms}{Interval between health checks}
}
\description{
Create a health check system
}
