% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{start_fun}
\alias{start_fun}
\title{Shiny Application starting function}
\usage{
start_fun(global_pars = TRUE)
}
\arguments{
\item{global_pars}{Set global parameters}
}
\description{
Shiny Application starting function
}
