% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_optimization.R
\name{stream_process_data}
\alias{stream_process_data}
\title{Data streaming utilities for large datasets}
\usage{
stream_process_data(data, chunk_size = 1000, process_chunk)
}
\arguments{
\item{data}{Large dataset}

\item{chunk_size}{Number of rows per chunk}

\item{process_chunk}{Function to process each chunk}
}
\description{
Data streaming utilities for large datasets
}
