% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tabs.R
\name{tab_menu}
\alias{tab_menu}
\title{Create a navigation menu}
\usage{
tab_menu(..., id = "")
}
\arguments{
\item{...}{Navigation menu items created with `navigation_menu_item()`}

\item{id}{HTML element ID}
}
\value{
a shiny tag
}
\description{
Creates a navigation menu that controls tab panels created with
`tabset_panel()`. The id of the menu and the panel must match.
}
