% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tabs.R
\name{tab_menu_item}
\alias{tab_menu_item}
\title{Create a navigation menu item}
\usage{
tab_menu_item(label = "", id = "", icon_svg = NULL)
}
\arguments{
\item{label}{Text to display}

\item{id}{unique tab-name}

\item{icon_svg}{icon for the menu}
}
\description{
Create a navigation menu item
}
\seealso{
tab_menu
}
