% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tabs.R
\name{tab_panel}
\alias{tab_panel}
\title{Create a panel}
\usage{
tab_panel(..., id = "")
}
\arguments{
\item{...}{content of the panel}

\item{id}{id of the panel. Must match the one created with `tab_menu_item()`}
}
\value{
a shiny.tag object
}
\description{
Create a panel
}
\seealso{
tabset_panel
}
