% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tabs.R
\name{tabset_panel}
\alias{tabset_panel}
\title{Create a tabset panel}
\usage{
tabset_panel(...)
}
\arguments{
\item{...}{panels created with `tab_panel()`}
}
\value{
a shiny.tag object
}
\description{
Creates a panel of tabs that is controled with `tab_menu()`
}
\seealso{
tab_menu
}
