% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_utils.R
\name{validate_numeric}
\alias{validate_numeric}
\title{Validate numeric data}
\usage{
validate_numeric(min_val = -Inf, max_val = Inf, allow_na = FALSE)
}
\arguments{
\item{min_val}{Minimum value}

\item{max_val}{Maximum value}

\item{allow_na}{Allow NA values}
}
\description{
Validate numeric data
}
