% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_horizontal_barplot}
\alias{plot_horizontal_barplot}
\title{Plot series using apexcharter}
\usage{
plot_horizontal_barplot(
  x_categories,
  series,
  y_formatter,
  x_formatter,
  type = "bar",
  sparkline = F,
  colors = NULL
)
}
\description{
Plot series using apexcharter
}
