% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot series using apexcharter}
\usage{
plot_timeseries(
  x_categories,
  series,
  y_formatter = V8::JS("function(x) {return x}"),
  type = "bar",
  sparkline = F,
  colors = NULL,
  stacked = F,
  enable_animations = FALSE
)
}
\arguments{
\item{x_categories}{X axis categories}

\item{series}{Time series in list format}

\item{y_formatter}{Y axis formatter}

\item{type}{Plot type}

\item{sparkline}{Enable sparkiline}

\item{colors}{Plot colors}

\item{stacked}{Enable stacked plot}
}
\value{
An apexcharter timeseries
}
\description{
Plot series using apexcharter
}
