% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries_apex}
\alias{plot_timeseries_apex}
\title{Plot series using apexcharter for more custom use}
\usage{
plot_timeseries_apex(
  data = NULL,
  type = "area",
  sparkline = F,
  y_formatter = NULL,
  plot_color = "#206bc4",
  lab_color = "#000000",
  mean = FALSE
)
}
\arguments{
\item{data}{Data to plot}

\item{type}{Plot type}

\item{sparkline}{Enable sparkiline}

\item{y_formatter}{Y axis formatter}

\item{plot_color}{Plot fill color}

\item{lab_color}{Plot lab color}

\item{mean}{Enable mean reference line}
}
\value{
An apexcharter timeseries
}
\description{
Plot series using apexcharter for more custom use
}
