% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_tabs.R
\name{tab_menu}
\alias{tab_menu}
\title{Create a navigation menu}
\usage{
tab_menu(...)
}
\arguments{
\item{...}{Navigation menu items created with `navigation_menu_item()`}
}
\value{
a shiny tag
}
\description{
Creates a navigation menu that controls tab panels created with
`tabset_panel()`. The id of the menu and the panel must match.
}
\examples{
tagList(
  tab_menu(
    tab_menu_item("First item", "tab-1"),
    tab_menu_item("Second item", "tab-2")
  ),
  tabset_panel(
    tab_panel(id = "tab-1", tags$h1("Overview")),
    tab_panel(id = "tab-2",tags$h1("Landings"))
  )
)
}
